Change Version to unity 2023.2.20.f1
Made Loading Screen for the Game
Fixed Font and UI that became a problem with the upgrade of the unity
Game works great
made couple of Shader Graphs shapes elements for making buttons and etc. to make the game more light instead of using textures. furthermore the shape will be more sharp then textures